broke up utils tests
