+++ propagator/tests/test_utils/test_gis.py
@@ -598,53 +598,6 @@ def test_aggregate_polygons():
     utils.cleanup_temp_results(testfile)
 
 
-def test_mask_array_with_flood():
-    zones = numpy.array([
-        [  1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   0],
-        [  1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   0],
-        [  1,   1,   1,   1,   1,   1,   0,   0,   0,   0,   0],
-        [  1,   1,   1,   1,   2,   2,   2,   2,   0,   0,   0],
-        [  0,   0,   0,   2,   2,   2,   2,   0,   0,   0,   0],
-        [  2,   2,   2,   2,   2,   2,   2,   0,   0,   0,   0],
-        [  2,   2,   2,   2,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-    ])
-
-    topo = numpy.array([
-        [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10.],
-        [ 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11.],
-        [ 2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12.],
-        [ 3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12., 13.],
-        [ 4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12., 13., 14.],
-        [ 5.,  6.,  7.,  8.,  9., 10., 11., 12., 13., 14., 15.],
-        [ 6.,  7.,  8.,  9., 10., 11., 12., 13., 14., 15., 16.],
-        [ 7.,  8.,  9., 10., 11., 12., 13., 14., 15., 16., 17.],
-        [ 8.,  9., 10., 11., 12., 13., 14., 15., 16., 17., 18.],
-        [ 9., 10., 11., 12., 13., 14., 15., 16., 17., 18., 19.],
-        [10., 11., 12., 13., 14., 15., 16., 17., 18., 19., 20.],
-    ])
-
-    known_flooded = numpy.array([
-        [  1,   1,   1,   1,   1,   1,   1,   0,   0,   0,   0],
-        [  1,   1,   1,   1,   1,   1,   0,   0,   0,   0,   0],
-        [  1,   1,   1,   1,   1,   0,   0,   0,   0,   0,   0],
-        [  1,   1,   1,   1,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  2,   2,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  2,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
-    ])
-
-    flooded = utils.flood_zones(zones, topo, 6.0)
-    nptest.assert_array_almost_equal(flooded, known_flooded)
-
-
 class Test_add_field_with_value(object):
     def setup(self):
         self.shapefile = resource_filename("propagator.testing.add_field_with_value", 'field_adder.shp')
