+++ propagator/tests/test_utils/test_misc.py
@@ -1,0 +1,52 @@
+import numpy
+
+import numpy.testing as nptest
+
+from propagator.utils import misc
+
+def test_mask_array_with_flood():
+    zones = numpy.array([
+        [  1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   0],
+        [  1,   1,   1,   1,   1,   1,   1,   1,   1,   1,   0],
+        [  1,   1,   1,   1,   1,   1,   0,   0,   0,   0,   0],
+        [  1,   1,   1,   1,   2,   2,   2,   2,   0,   0,   0],
+        [  0,   0,   0,   2,   2,   2,   2,   0,   0,   0,   0],
+        [  2,   2,   2,   2,   2,   2,   2,   0,   0,   0,   0],
+        [  2,   2,   2,   2,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+    ])
+
+    topo = numpy.array([
+        [ 0.,  1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10.],
+        [ 1.,  2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11.],
+        [ 2.,  3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12.],
+        [ 3.,  4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12., 13.],
+        [ 4.,  5.,  6.,  7.,  8.,  9., 10., 11., 12., 13., 14.],
+        [ 5.,  6.,  7.,  8.,  9., 10., 11., 12., 13., 14., 15.],
+        [ 6.,  7.,  8.,  9., 10., 11., 12., 13., 14., 15., 16.],
+        [ 7.,  8.,  9., 10., 11., 12., 13., 14., 15., 16., 17.],
+        [ 8.,  9., 10., 11., 12., 13., 14., 15., 16., 17., 18.],
+        [ 9., 10., 11., 12., 13., 14., 15., 16., 17., 18., 19.],
+        [10., 11., 12., 13., 14., 15., 16., 17., 18., 19., 20.],
+    ])
+
+    known_flooded = numpy.array([
+        [  1,   1,   1,   1,   1,   1,   1,   0,   0,   0,   0],
+        [  1,   1,   1,   1,   1,   1,   0,   0,   0,   0,   0],
+        [  1,   1,   1,   1,   1,   0,   0,   0,   0,   0,   0],
+        [  1,   1,   1,   1,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  2,   2,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  2,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+        [  0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0],
+    ])
+
+    flooded = misc.flood_zones(zones, topo, 6.0)
+    nptest.assert_array_almost_equal(flooded, known_flooded)
+
